Performance improvements (#340)

+++ src/mflux/models/common/schedulers/linear_scheduler.py
@@ -51,5 +51,5 @@ class LinearScheduler(BaseScheduler):
         return timesteps
 
     def step(self, noise: mx.array, timestep: int, latents: mx.array, **kwargs) -> mx.array:
-        dt = self._sigmas[timestep + 1] - self._sigmas[timestep]
-        return latents + noise * dt
+        dt = (self._sigmas[timestep + 1] - self._sigmas[timestep]).astype(latents.dtype)
+        return latents + noise.astype(latents.dtype) * dt
